refactor validateSpanOptions into span factory and remove logger param from platformExtensions function
