+++ packages/core/lib/span-factory.ts
@@ -5,10 +5,10 @@ import { type Configuration, type Logger } from './config'
 import { type IdGenerator } from './id-generator'
 import { type Processor } from './processor'
 import { type ReadonlySampler } from './sampler'
-import { SpanInternal, type Span, type SpanOptions } from './span'
+import { SpanInternal, type Span, type SpanOptions, type SpanOptionSchema, type InternalSpanOptions, coreSpanOptionSchema } from './span'
 import { type SpanContextStorage } from './span-context'
 import { timeToNumber } from './time'
-import { isSpanContext } from './validation'
+import { isObject, isSpanContext } from './validation'
 
 export class SpanFactory <C extends Configuration> {
   private processor: Processor
@@ -129,5 +129,39 @@ export class SpanFactory <C extends Configuration> {
         this.endSpan(span, safeEndTime)
       }
     }
+  }
+
+  validateSpanOptions<O extends SpanOptions> (name: string, options: unknown, schema: SpanOptionSchema = coreSpanOptionSchema): InternalSpanOptions<O> {
+    let warnings = ''
+    const cleanOptions: Record<string, unknown> = {}
+
+    if (typeof name !== 'string') {
+      warnings += `\n  - name should be a string, got ${typeof name}`
+      name = String(name)
+    }
+
+    if (options !== undefined && !isObject(options)) {
+      warnings += '\n  - options is not an object'
+    } else {
+      const spanOptions = options || {}
+      for (const option of Object.keys(schema)) {
+        if (Object.prototype.hasOwnProperty.call(spanOptions, option) && spanOptions[option] !== undefined) {
+          if (schema[option].validate(spanOptions[option])) {
+            cleanOptions[option] = spanOptions[option]
+          } else {
+            warnings += `\n  - ${option} ${schema[option].message}, got ${typeof spanOptions[option]}`
+            cleanOptions[option] = schema[option].getDefaultValue(spanOptions[option])
+          }
+        } else {
+          cleanOptions[option] = schema[option].getDefaultValue(spanOptions[option])
+        }
+      }
+    }
+
+    if (warnings.length > 0) {
+      this.logger.warn(`Invalid span options${warnings}`)
+    }
+
+    return { name, options: cleanOptions } as unknown as InternalSpanOptions<O>
   }
 }
