+++ packages/core/lib/span.ts
@@ -166,37 +166,3 @@ export const coreSpanOptionSchema: SpanOptionSchema = {
     validate: isBoolean
   }
 }
-
-export function validateSpanOptions<O extends SpanOptions> (name: string, options: unknown, schema: SpanOptionSchema, logger: Logger): InternalSpanOptions<O> {
-  let warnings = ''
-  const cleanOptions: Record<string, unknown> = {}
-
-  if (typeof name !== 'string') {
-    warnings += `\n  - name should be a string, got ${typeof name}`
-    name = String(name)
-  }
-
-  if (options !== undefined && !isObject(options)) {
-    warnings += '\n  - options is not an object'
-  } else {
-    const spanOptions = options || {}
-    for (const option of Object.keys(schema)) {
-      if (Object.prototype.hasOwnProperty.call(spanOptions, option) && spanOptions[option] !== undefined) {
-        if (schema[option].validate(spanOptions[option])) {
-          cleanOptions[option] = spanOptions[option]
-        } else {
-          warnings += `\n  - ${option} ${schema[option].message}, got ${typeof spanOptions[option]}`
-          cleanOptions[option] = schema[option].getDefaultValue(spanOptions[option])
-        }
-      } else {
-        cleanOptions[option] = schema[option].getDefaultValue(spanOptions[option])
-      }
-    }
-  }
-
-  if (warnings.length > 0) {
-    logger.warn(`Invalid span options${warnings}`)
-  }
-
-  return { name, options: cleanOptions } as unknown as InternalSpanOptions<O>
-}
